mz-template-fastapi/mz-template-fastapi/requirements.txt
#configure 1.0.0 above version from libraries
aiofiles
alembic
anyio
asyncpg
azure-storage-blob
azure-storage-queue
celery
fastapi
fastapi-pagination
fastapi-security
flower
humanize
langid
langdetect
openpyxl
opentelemetry-api
opentelemetry-instrumentation-fastapi
opentelemetry-sdk
pandas
pgvector
psycopg2
psycopg2-binary
pydantic-settings
pydantic==2.8.2
python-dotenv
python-jose[cryptography]
python-json-logger
python-multipart
requests
rsa
semantic-kernel
sqlalchemy
typing_extensions
uuid
uvicorn
xlsxwriter